+++ packages/shared/src/react/hooks/index.ts
@@ -15,3 +15,13 @@ export { usePaymentMethods as __experimental_usePaymentMethods } from './usePaym
 export { usePlans as __experimental_usePlans } from './usePlans';
 export { useSubscription as __experimental_useSubscription } from './useSubscription';
 export { useCheckout as __experimental_useCheckout } from './useCheckout';
+
+/**
+ * Internal hooks to be consumed only by `@clerk/clerk-js`.
+ * These are not considered part of the public API and their query keys can change without notice.
+ *
+ * These exist here in order to keep RQ and SWR implementations in a centralized place.
+ */
+export { __internal_useStatementQuery } from './useStatementQuery';
+export { __internal_usePlanDetailsQuery } from './usePlanDetailsQuery';
+export { __internal_usePaymentAttemptQuery } from './usePaymentAttemptQuery';